requirements.txt
# Core ML and NLP libraries
torch>=2.1.0
transformers>=4.35.0
datasets>=2.14.0
scikit-learn>=1.3.2
numpy>=1.24.3
pandas>=2.1.2

# Deep Learning and NLP specific
sentencepiece>=0.1.99
tokenizers>=0.14.1
nltk>=3.8.1
spacy>=3.7.1
networkx>=3.2.1  # For knowledge graphs

# API and Web Framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
starlette>=0.31.1

# Database and Caching
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1
redis>=5.0.1

# Data Processing and Visualization
jupyter>=1.0.0
matplotlib>=3.8.1
seaborn>=0.13.0
plotly>=5.18.0

# API Clients and External Services
yfinance>=0.2.31  # For financial data
alpha_vantage>=2.3.1  # Additional financial data
requests>=2.31.0
aiohttp>=3.9.0

# Testing and Monitoring
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.1
prometheus-client>=0.18.0
grafana-api>=1.0.3

# Utilities and Logging
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.1
loguru>=0.7.2
mlflow>=2.8.0  # For experiment tracking
wandb>=0.15.12  # For experiment tracking

# Security and Authentication
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
authlib>=1.2.1
cryptography>=41.0.4

# Deployment and Infrastructure
docker>=6.1.3
kubernetes>=27.2.0
gunicorn>=21.2.0
supervisor>=4.2.5
